Add page with js module done
