+++ modules/catalog.js
@@ -1,17 +1,11 @@
-/* eslint-disable max-classes-per-file */
+import Book from './book.js';
 
-import {form, display, catalog, infoDisplayS } from '../index.js';
+const form = document.getElementById('frm_books');
+const display = document.getElementById('display');
+const infoDisplayS = document.querySelector('small');
 let i = 0;
 
-export class Book {
-  constructor(id, title, author) {
-    this.id = id;
-    this.title = title;
-    this.author = author;
-  }
-}
-
-export class Catalog {
+export default class Catalog {
   constructor() {
     this.books = [];
   }
@@ -31,8 +25,8 @@ export class Catalog {
         const button = event.target;
         const parent = button.parentNode;
         parent.parentNode.removeChild(parent);
+        //console.log(this.books[idx]);
         this.removeBook(idx);
-        console.log(idx);
         this.saveBooks();
         if (i > 0) {
           i -= 1;
@@ -65,7 +59,8 @@ export class Catalog {
       book.author = form.elements.author.value;
       form.elements.title.value = '';
       form.elements.author.value = '';
-      catalog.addBook(book);
+      console.log(book);
+      this.addBook(book);
       this.displayData(i);
       this.saveBooks();
       infoDisplayS.innerHTML = 'Book added successfully!';
@@ -75,17 +70,21 @@ export class Catalog {
   }
 
   saveBooks() {
-    const strData = JSON.stringify(this.books);  
+    const strData = JSON.stringify(this.books);
     localStorage.setItem('data', strData);
   }
 
   restoreBooks() {
     let data = localStorage.getItem('data');
     if (!data) {
-      data = {};
+      data = [];
     } else {
       const objData = JSON.parse(data);
-      this.books = objData;
+      if (Array.isArray(objData)) {
+        this.books = objData;
+      } else {
+        console.error('Data retrieved from local storage is not an array!');
+      }
     }
   }
